requirements.txt
feedparser>=6.0
openai>=1.0
pyyaml>=6.0
python-dateutil>=2.8
pytest>=7.0
